Mochi cuda Docker file now builds Sonata with libfabric TCP support Provenance DB connects now via tcp. Added option to provider to specify ip address/port Replaced threading in provenance DB with Thallium calls to prevent interference between pthread and argobots, as advised by Mochi team Provenance DB integration tests now run by default if provenance DB was built

+++ test/run_provdb_client_test.sh
@@ -1,11 +1,18 @@
 #!/bin/bash
 
-rm provdb.unqlite  provider.address
+if [ -f "../bin/provdb_admin" ]; then
+    rm provdb.unqlite  provider.address
 
-../bin/provdb_admin &
-admin=$!
-sleep 1
+    ip=$(hostname -i)
+    port=1234
 
-./mainProvDBclient $(cat provider.address)
+    ../bin/provdb_admin ${ip}:${port} &
+    admin=$!
+    sleep 1
 
-kill $admin
+    ./mainProvDBclient $(cat provider.address)
+
+    kill $admin
+else 
+    echo "Provenance DB was not built"
+fi
